+++ docs/examples/scripts/generate_figures.py
@@ -15,18 +15,18 @@ single_example_fig = context._figures["source_w1a_D1000B340"]
 single_example_fig.savefig(FIGURES_PATH / "single_example.png")
 
 folder_path = EXAMPLES_PATH / "data/folder"
-folder_conig_path = EXAMPLES_PATH / "folder_example_config.yaml"
-context = run(folder_conig_path, folder_path)
-folder_gain_fig = context._figures["gain"]
+folder_config_path = EXAMPLES_PATH / "folder_example_config.yaml"
+context = run(folder_config_path, folder_path)
+folder_gain_fig = context.folder_ctx("folder")._figures["gain"]
 folder_gain_fig.savefig(FIGURES_PATH / "folder_gain.png")
-folder_resolution_fig = context._figures["resolution"]
+folder_resolution_fig = context.folder_ctx("folder")._figures["resolution"]
 folder_resolution_fig.savefig(FIGURES_PATH / "folder_resolution.png")
 
 plt.close("all")
 gain_trend_path = EXAMPLES_PATH / "data/gain_trend"
 gain_trend_config_path = EXAMPLES_PATH / "gain_trend_example_config.yaml"
 context = run(gain_trend_config_path, gain_trend_path)
-gain_trend_fig = context._figures["gain"]
+gain_trend_fig = context.folder_ctx("gain_trend")._figures["gain"]
 gain_trend_fig.savefig(FIGURES_PATH / "gain_trend.png")
 
 folder1_path = EXAMPLES_PATH / "data/folder1"