+++ docs/release_notes.md
@@ -3,8 +3,11 @@
 
 
 - Analysis output file refactored.
+- Compare tasks refactored into a single compare task that can handle all the different quantities to compare.
 - Pull requests merged and issues closed:
     - [PR #60](https://github.com/augustocattafesta/muGPD/pull/60)
+    - [PR #62](https://github.com/augustocattafesta/muGPD/pull/62)
+    - [Issue #49](https://github.com/augustocattafesta/muGPD/issues/61)
 
 ### Version 0.0.2 (2026-03-20)